Change Current value for CurrentImage method
To improve performance, the full image should only be returned/drawn
when the image is transformed. For frames, just the next frame should
be drawn. This commit temporarily hurts performance by re-combining
every frame for each frame, but is made in preparation for the
CurrentImage method to only be used sparingly.