Track width/height instead of using resized image
Instead of *always* resizing an image, and using that resized image to
get the appropriate bounds, simply stores the width/height of the
current image/animation frame in two variables. This fixes an issue with
zooming sometimes causing a freeze.

Still not entirely sure *why* zooming in/out could cause a freeze
:shrug:
It seems to caused by treating the `gif.Helper` like any other image and
trying to resize it, which would call `CurrentImage`. Perhaps attempting
a resize of an image that's asynchronously updating, and therefore
potentially *changing during the resize event*, could cause things to
freeze.

Fixes #38