Buffer zoom channel
Deadlock caused by main loop sending to zoom channel and waiting for it
to be read while animation loop is sending to frame channel and waiting
for it to be read. By buffering the zoom channel, the main loop can send
to it without waiting for it to be read.

Fixes #45