Use goroutine when resizing animation
There is no need for the main loop to be blocked until the animator
function reads the new zoom value, so a goroutine is used to make the
send non-blocking. This can prevent a deadlock caused by trying to send
a new zoom value while the animator is trying to send a new animation
frame.

Because sending on the zoom channel (should) always be non-blocking now,
the zoom channel no longer needs to be buffered.

Fixes #54