Support replace (`DisposalBackground`)
Instead of combining on *always*, frames will be combined only if
disposal method is none or if it's an unknown/unhandled disposal method.
If the disposal method is `DisposalBackground`, draws the frame over a
transparent background.

Despite the code being `DisposalBackground`, and a background color
being available, modern browsers and image-viewing tools ignore the
background color and instead use transparent as the only background
color. To match that behavior, the response to `DisposalBackground` is
to do nothing, and simply draw the next frame over an empty image.

Resolves #35