Add test for redrawing title
Bug #66 is caused by the title channel simply setting a `title` variable
instead of triggering a title to be drawn. This can result in a race
condition where the image channel has been read before the title
channel, resulting in the title being a "zero value" (`""` for strings)
when the image is drawn.

This adds a test to ensure that drawing a title clears the previous
title.