Draw image and title separately
Instead of title always relying on the image to be re-drawn in order to
also be re-drawn, the title can now trigger its own drawing when it is
received on a channel. This is achieved by the title clearing and
re-drawing the title rows, and the image clearing and re-drawing the
image rows, sans title rows.

This has the added benefit of rendering the title before a large image
has finished processing, allowing the user to see that the next image is
being loaded.

Fixes #66